making logic more simpler
